+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## Added
+- editor interface customization
+
 ## 1.0.0
 ## Changed
 - fix saving intermediate versions when editing (forcesave)